LayoutTests/crypto/workers/crypto-random-values-worker-expected.txt

[Worker] Tests crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/crypto-random-values-worker.js
PASS [Worker] 'crypto' in self is true
PASS [Worker] 'getRandomValues' in self.crypto is true
PASS [Worker] matchingBytes < 100 is true
PASS successfullyParsed is true

TEST COMPLETE
